Add files via upload
This is a project on Stock Market Analysis And Forecasting Using Deep Learning. Here we use python, pandas, matplotlib, numpy, plotly, pytorch to implement our model.

A stock market, equity market, or share market is the aggregation of buyers and sellers of stocks (also called shares), which represent ownership claims on businesses; these may include securities listed on a public stock exchange, as well as stock that is only traded privately, such as shares of private companies which are sold to investors through equity crowdfunding platforms. Investment in the stock market is most often done via stock brokerages and electronic trading platforms. Investment is usually made with an investment strategy in mind.

The task of stock prediction has always been a challenging problem for statistics experts. The main reason behind this prediction is buying stocks that are likely to increase in price and then selling stocks that are probably to fall. Generally, there are two ways for stock market prediction. Fundamental analysis relies on a company’s technique and fundamental information like market position, expenses, and annual growth rates. The second one is the technical analysis method, which concentrates on previous stock prices and values.

We can see that the stock market is a profitable resource for a person but there are some risk factors too. The share market is continuously getting ups and downs in this field. So we have to be very conscious about the market price and the stock increment & decrement factor. For that, we need a broker, who has a strong acquaintance with the share market policy.

But after the evolution of data science, deep learning, and time series analysis the task of a stock buyer has become comprehensively easy. He or she can easily search in Google and get the necessary information of stock market policy for the felicity of data science.

In the first part of our project, we will try to analyze the data. and in the second part, we will forecast the stock market.

Dataset:
Here we will use multiple stock market datasets such as
Google(2006–2018)
Microsoft(2006–2018)
IBM(2006–2018)
Amazon(2006–2018)

Stock Market Analysis:
We will find the followings:-
Describe
The distribution of close and open.
Correlation between close and open.
Visualize the attributes[Open, High, Low, Close, volume] of our datasets.
Compare the “High” and “Close” of each dataset.
At last, the trend and seasonality in the dataset.